webapp/allergenicity-webapp-streamlit/requirements.txt
streamlit
torch
numpy
joblib
shap
fair-esm
matplotlib
xgboost